+++ src/api/updateChecking.tsx
@@ -2,6 +2,12 @@ import EquationCitator from "@/main";
 import { Notice, requestUrl } from "obsidian";
 const versionUrl = "https://raw.githubusercontent.com/FRIEDparrot/obsidian-equation-citator/refs/heads/master/manifest.json"
 
+
+/**
+ * Check if a new version of the plugin is available by fetching the latest version from GitHub and comparing it with the current version.
+ * 
+ * This is never called now and just retained for possibly future use.
+ */
 export async function isUpdateAvailable(plugin: EquationCitator, use_notice=true): Promise<boolean> {
     try {
         const response = await requestUrl(versionUrl);